replaced obsolete Tools::toString(bool/double/float/int) with template version Tools::toString(const T&)

+++ branches/AST_Bugfix/Projects/Core/Core/VirtualMachine/StackFrame.h
@@ -59,7 +59,7 @@ private:
 	void operator=(const StackFrame&)/* = delete*/;
 
 private:
-	unsigned long mLevel;
+	unsigned int mLevel;
 	ParameterList mParameters;
 	Runtime::Object mReturnValue;
 	IScope* mScope;
